requirements.txt
rich
websockets
python-dotenv
